track equipment usage by temperature controller
