+++ controller/temperature/controller.go
@@ -11,16 +11,6 @@ import (
 
 const Bucket = "temperature"
 
-type Measurement struct {
-	Time        string  `json:"time"`
-	Temperature float32 `json:"temperature"`
-}
-
-type Usage struct {
-	Heater int `json:"heater"`
-	Cooler int `json:"cooler"`
-}
-
 type Controller struct {
 	config     Config
 	stopCh     chan struct{}
@@ -135,86 +125,6 @@ func (c *Controller) Stop() {
 	c.stopCh <- struct{}{}
 }
 
-func (c *Controller) switchHeater(on bool) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if on != c.heater.On {
-		log.Println("Temperature subsystem - switching heater on:", on)
-		c.heater.On = on
-		if err := c.equipments.Update(c.heater.ID, *c.heater); err != nil {
-			c.heater.On = !on
-			return err
-		}
-	}
-	c.telemetry.EmitMetric("heater", on)
-	return nil
-}
-
-func (c *Controller) switchCooler(on bool) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if on != c.cooler.On {
-		log.Println("Temperature subsystem - switching cooler on:", on)
-		c.cooler.On = on
-		if err := c.equipments.Update(c.cooler.ID, *c.cooler); err != nil {
-			c.cooler.On = !on
-			return err
-		}
-	}
-	c.telemetry.EmitMetric("cooler", on)
-	return nil
-}
-
-func (c *Controller) warmUp() error {
-	if c.cooler == nil {
-		log.Println("Temperature subsystem: WarmUp routine. cooler is not set.")
-	} else {
-		if c.cooler.On {
-			log.Println("WARNING: Possible flapping. Turning off cooler due to warm up routine.")
-			if err := c.switchCooler(false); err != nil {
-				return err
-			}
-		}
-	}
-	if c.heater == nil {
-		log.Println("Temperature subsystem: WarmUp routine. heater is not set")
-	} else {
-		if c.heater.On {
-			log.Println("Temperature subsystem: Heater is already on. Skipping switch on")
-			return nil
-		}
-		if err := c.switchHeater(true); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *Controller) coolDown() error {
-	if c.heater == nil {
-		log.Println("Temperature subsystem: CoolDown routine. heater is not set")
-	} else {
-		if c.heater.On {
-			log.Println("WARNING: Temperature subsystem: Possible flapping. Turning off heater due to cool down routine.")
-			if err := c.switchHeater(false); err != nil {
-				return err
-			}
-		}
-	}
-	if c.cooler == nil {
-		log.Println("Temperature subsystem: CoolDown routine. cooler is not set")
-	} else {
-		if c.cooler.On {
-			log.Println("Temperature subsystem: Cooler is already on. Skipping switch on")
-			return nil
-		}
-		if err := c.switchCooler(true); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 func (c *Controller) control(reading float32) error {
 	if err := c.loadEquipments(); err != nil {
 		log.Println("ERROR: temperature subsystem. Failed to load equipments: Error:", err)
@@ -231,13 +141,4 @@ func (c *Controller) control(reading float32) error {
 		c.switchOffAll()
 	}
 	return nil
-}
-
-func (c *Controller) switchOffAll() {
-	if (c.cooler != nil) && c.cooler.On {
-		c.switchCooler(false)
-	}
-	if (c.heater != nil) && c.heater.On {
-		c.switchHeater(false)
-	}
 }
